requirements.txt
# 不加會報錯

# django
django==4.0.3
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-bulk-update==2.2.0
django-crontab==0.7.1
django-debug-toolbar==3.6.0

# 跨網域call api
django-cors-headers==3.13.0

# 資料庫
mysqlclient==2.1.0
PyMySQL==1.0.2

# 資料處理
pandas==1.4.1
openpyxl==3.0.9
cn2an==0.5.16 # 中文數字轉阿拉伯數字

# API 框架
djangorestframework==3.13.1
drf-spectacular==0.22.0

# 影像處理
Pillow==9.0.1

# 座標處理
django-leaflet==0.28.3
shapely==1.8.1.post1
scipy==1.8.0

# 進度條
tqdm==4.64.0
rich==12.6.0

# server
uwsgi==2.0.20
uwsgitop

# cache
django-redis==5.2.0
redis==4.3.3

# celery
celery==5.2.7
django_celery_beat==2.3.0
flower==1.0.0

python-decouple==3.6

# pdf 製作
xhtml2pdf

# 拆解正規化地址用
zipcodetw

# 圖表
matplotlib==3.6.2

# common
--trusted-host diablo.yeshome.net.tw --extra-index-url http://diablo.yeshome.net.tw:25386
wsos_common>=0.2.5

